Added picture decoder in C
The first line of the inputted code is a key. Each line after must be exclusive-ored with the key to create a new line. The 1s in the new line are replaced with an X, and the 0s in the new line are replaces with a space to create a photo. 

Example Input:

01010100011000
01011011101000
01001011100110
11001011100111
10101011111111
10101011100111
00101011111110
10101101100111
10110000111111
01000110001010
01010001010010
01010101111110


Example Output (cat):

      XX             XX 
      X  X         X  X 
      X   X       X   X 
      X  XXXXXX  X 
    XXXXXXXXXXXX 
    XXXX  XX  XXXX
  XXXXXXXXXXXXXX
   XXXXXX  XXXXXX
    XXXXXXXXXXXX 
      XXXXXXXXXX  
          XXXXXX